requirements.txt
flask
line-bot-sdk
pandas
requests
openai
python-dotenv
gunicorn